+++ src/OPI/opi_propagator.h
@@ -88,7 +88,7 @@ namespace OPI
              * @param dt The time step, in seconds, from last propagation.
              * @return OPI::SUCCESS if propagation was successful, or other error code.
              */
-            ErrorCode propagate(Population& data, double julian_day, float dt);
+            ErrorCode propagate(Population& data, double julian_day, double dt);
 
             /**
              * @brief propagate Starts the index-based propagation for the given time step.
@@ -108,7 +108,7 @@ namespace OPI
              * was performed with OPI's inherent method (which should still give you valid results); or
              * any other error code returned by the plugin.
              */
-			ErrorCode propagate(Population& data, IndexList& indices, double julian_day, float dt);
+            ErrorCode propagate(Population& data, IndexList& indices, double julian_day, double dt);
 
             /**
              * @brief propagate Starts propagation with individual times for each object.
@@ -127,7 +127,7 @@ namespace OPI
              * was performed with OPI's inherent method (which should still give you valid results); or
              * any other ErrorCode returned by the plugin.
              */
-            ErrorCode propagate(Population& data, double* julian_days, int length, float dt);
+            ErrorCode propagate(Population& data, double* julian_days, int length, double dt);
 
 			//! Assigns a module to this propagator
 			/**
@@ -188,10 +188,10 @@ namespace OPI
 			void useModules();
 			//! The actual propagation implementation
 			//! The C Namespace equivalent for this function is OPI_Plugin_propagate
-			virtual ErrorCode runPropagation(Population& data, double julian_day, float dt ) = 0;
+            virtual ErrorCode runPropagation(Population& data, double julian_day, double dt) = 0;
 			//! Override this to implement an index-based propagation
 			//! The C Namespace equivalent for this function is OPI_Plugin_propagateIndexed
-			virtual ErrorCode runIndexedPropagation(Population& data, IndexList& indices, double julian_day, float dt );
+            virtual ErrorCode runIndexedPropagation(Population& data, IndexList& indices, double julian_day, double dt);
             //! Override this to implement propagation with individual times.
             //! OPI will make sure that the julian_days vector length matches that of the Population.
             virtual ErrorCode runMultiTimePropagation(Population& data, double* julian_days, float dt);
